3.4-django-testing/django_testing/requirements.txt
django~=5.1.4
djangorestframework~=3.15.2
django-filter~=24.3
psycopg2-binary

pytest~=8.3.4
model-bakery~=1.20.0